glucose-hack-track: get diff down to 1000
As the competition rules ask for small diffs, we need to define a few
things. Most importantly, the for loop to print unsat proofs for clauses.

Furthermore, some conditions are negated to not use continue, as well as
logical or (||) is replaced with the plus operator (+).

With this patch, the diff is exactly 1000 chars.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>